Modificar métodos obtenerVuelosDisponibles y reservarVuelo. Adaptar entidades y dto para dichos métodos. Modificar data.sql.

+++ src/main/java/com/codoacodo/flysky/demo/dto/response/VueloReservaDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @EqualsAndHashCode
 public class VueloReservaDto {
 
-    private Integer numeroVuelo;
+    private Long numeroVuelo;
     private String aerolinea;
     private LocalDateTime fechaHoraPartida;
     private LocalDateTime fechaHoraLlegada;
